forcing connection method testing
the transport field in device config is the one that seems to force a connection. 1 for USB 2(?) for wireless. but for some reason calling setTransport with 1 results in the actual value having some crazy value. The assembly for this function is a simple assignment. Not sure what's going on with that. Forcing the value to one through the debugger does seem to actually force the connection method. The annoying thing is that once it's forced, that connection method becomes sticky. meaning, once USB is forced, disconnecting the USB cable does not cause the connection to fall back to wireless.